fix: image s3 upload로 변경

+++ server/API/validation/user.ts
@@ -3,12 +3,16 @@ import { CustomRequest } from '../../type/type';
 
 const dotenv: any = require('dotenv');
 dotenv.config();
-
+const multer: any = require('multer');
+const multerS3: any = require('multer-s3');
+const aws: any = require('aws-sdk');
+aws.config.loadFromPath(__dirname + '/../../aws-config.json');
+const s3 = new aws.S3();
 const db: any = require('../../models/index');
 const jwt: any = require('jsonwebtoken');
 const bcrypt: any = require('bcrypt');
 const { disconnectKakao } = require('../../method/oauth');
-
+const method: any = require('../../method/custom');
 import accessTokenRequestValidation from '../../method/token';
 
 interface UserValidation {
@@ -46,6 +50,7 @@ interface UserValidation {
     res: Response,
     next: NextFunction,
   ): Promise<any>;
+  uploadImage(req: Request, res: Response, next: NextFunction): any;
 }
 
 const userValidation: UserValidation = {
@@ -305,15 +310,22 @@ const userValidation: UserValidation = {
     const userInfo: any = await db['User'].findOne({
       where: { email },
     });
+    if (!userInfo) {
+      req.sendData = { message: 'no exists userInfo' };
+      next();
+      return;
+    }
     if (social_login === 'kakao') {
-      const kakaoId = await disconnectKakao(db['User'].auth_code);
+      const kakaoId = await disconnectKakao(userInfo.auth_code);
       if (kakaoId) {
         db['User'].destroy({
           where: { email: userInfo.email },
         });
         req.sendData = { message: 'ok' };
+        next();
       } else {
         req.sendData = { message: 'err' };
+        next();
       }
     } else if (social_login === 'google') {
       db['User'].destroy({
@@ -322,7 +334,7 @@ const userValidation: UserValidation = {
       req.sendData = { message: 'ok' };
       next();
     } else if (social_login === 'original') {
-      const { email, password } = req.body;
+      const { password } = req.body;
       bcrypt.compare(
         password,
         userInfo.password,
@@ -586,6 +598,7 @@ const userValidation: UserValidation = {
     }
     const { email } = req.params;
     const { filename } = req.file;
+
     db['User']
       .update(
         {
@@ -638,6 +651,58 @@ const userValidation: UserValidation = {
       req.sendData = { message: 'insufficient parameters supplied' };
       next();
     }
+  },
+
+  uploadImage(req: Request, res: Response, next: NextFunction): any {
+    const DIR_NAME = __dirname + '/../..';
+    // const storage: any = multer.diskStorage({
+    //   destination: (req: any, file: any, cb: any): void => {
+    //     cb(null, DIR_NAME + '/image/user/'); // 파일 업로드 경로
+    //   },
+    //   filename: (req: any, file: any, cb: any): void => {
+    //     const regex: any = /^[a-z|A-Z|0-9|]+$/;
+    //     let dot =
+    //       file.originalname.split('.')[file.originalname.split('.').length - 1];
+    //     if (dot !== 'png' && dot !== 'jpg' && dot !== 'jepg') {
+    //       // return cb(new Error('Only .png, .jpg and .jpeg format allowed!'));
+    //       res
+    //         .status(400)
+    //         .send({ message: 'Only .png, .jpg and .jpeg format allowed' });
+    //       return;
+    //     }
+    //     let name = file.originalname;
+    //     if (!regex.test(name)) {
+    //       name = Math.random().toString(36).substring(0, 8) + '.' + dot;
+    //     }
+    //     cb(null, method.randomString(8, name)); //파일 이름 설정
+    //   },
+    // });
+    const storage: any = multerS3({
+      s3: s3,
+      bucket: 'rightnow-image',
+      acl: 'public-read',
+      key: (req: any, file: any, cb: any) => {
+        const regex: any = /^[a-z|A-Z|0-9|]+$/;
+        let dot =
+          file.originalname.split('.')[file.originalname.split('.').length - 1];
+        if (dot !== 'png' && dot !== 'jpg' && dot !== 'jepg') {
+          res
+            .status(400)
+            .send({ message: 'Only .png, .jpg and .jpeg format allowed' });
+          return;
+        }
+        let name = file.originalname;
+        if (!regex.test(name)) {
+          name = Math.random().toString(36).substring(0, 8) + '.' + dot;
+        }
+        cb(null, 'user/' + method.randomString(8, name));
+      },
+    });
+    let upload: any = multer({
+      storage,
+      limits: { fileSize: 1000 * 1000 * 10 },
+    });
+    upload.single('file')(req, res, next);
   },
 };
